+++ lib/karafka/connection/client.rb
@@ -36,6 +36,12 @@ module Karafka
         # Marks if we need to offset. If we did not store offsets, we should not commit the offset
         # position as it will crash rdkafka
         @offsetting = false
+        # We need to keep track of what we have paused for resuming
+        # In case we loose partition, we still need to resume it, otherwise it won't be fetched
+        # again if we get reassigned to it later on. We need to keep them as after revocation we
+        # no longer may be able to fetch them from Kafka. We could build them but it is easier
+        # to just keep them here and use if needed when cannot be obtained
+        @paused_tpls = Hash.new { |h, k| h[k] = {} }
       end
 
       # Fetches messages within boundaries defined by the settings (time, size, topics, etc).
@@ -148,6 +154,8 @@ module Karafka
 
         return unless tpl
 
+        @paused_tpls[topic][partition] = tpl
+
         @kafka.pause(tpl)
 
         @kafka.seek(pause_msg)
@@ -171,6 +179,14 @@ module Karafka
 
         tpl = topic_partition_list(topic, partition)
 
+        # If we were able to get the tpc, we should cache if for future usage
+        if tpl
+          @paused_tpls[topic][partition] = tpl
+        else
+          # If we were not able, let's try to reuse the one we have (if we have)
+          tpl ||= @paused_tpls[topic][partition]
+        end
+
         return unless tpl
 
         @kafka.resume(tpl)
@@ -190,6 +206,7 @@ module Karafka
       # Marks given message as consumed.
       #
       # @param [Karafka::Messages::Message] message that we want to mark as processed
+      # @return [Boolean] true if successful. False if we no longer own given partition
       # @note This method won't trigger automatic offsets commits, rather relying on the offset
       #   check-pointing trigger that happens with each batch processed
       def mark_as_consumed(message)
@@ -199,8 +216,10 @@ module Karafka
       # Marks a given message as consumed and commits the offsets in a blocking way.
       #
       # @param [Karafka::Messages::Message] message that we want to mark as processed
+      # @return [Boolean] true if successful. False if we no longer own given partition
       def mark_as_consumed!(message)
-        mark_as_consumed(message)
+        return false unless mark_as_consumed(message)
+
         commit_offsets!
       end
 
@@ -211,28 +230,42 @@ module Karafka
         @mutex.synchronize do
           @closed = false
           @offsetting = false
+          @paused_tpls.clear
           @kafka = build_consumer
         end
       end
 
       private
 
+      # When we cannot store an offset, it means we no longer own the partition
+      #
       # Non thread-safe offset storing method
       # @param message [Karafka::Messages::Message]
+      # @return [Boolean] true if we could store the offset (if we still own the partition)
       def internal_store_offset(message)
         @offsetting = true
         @kafka.store_offset(message)
+        true
+      rescue Rdkafka::RdkafkaError => e
+        return false if e.code == :assignment_lost
+        return false if e.code == :state
+
+        raise e
       end
 
       # Non thread-safe message committing method
       # @param async [Boolean] should the commit happen async or sync (async by default)
+      # @return [Boolean] true if offset commit worked, false if we've lost the assignment
       def internal_commit_offsets(async: true)
-        return unless @offsetting
+        return true unless @offsetting
 
         @kafka.commit(nil, async)
         @offsetting = false
+
+        true
       rescue Rdkafka::RdkafkaError => e
-        return if e.code == :no_offset
+        return false if e.code == :assignment_lost
+        return false if e.code == :no_offset
 
         raise e
       end
@@ -250,7 +283,8 @@ module Karafka
 
           @kafka.close
           @buffer.clear
-          @rebalance_manager.clear
+          # @note We do not clear rebalance manager here as we may still have revocation info here
+          # that we want to consider valid prior to running another reconnection
         end
       end
 
@@ -303,7 +337,13 @@ module Karafka
 
         time_poll.backoff
 
-        retry
+        # We return nil, so we do not restart until running the whole loop
+        # This allows us to run revocation jobs and other things and we will pick up new work
+        # next time after dispatching all the things that are needed
+        #
+        # If we would retry here, the client reset would become transparent and we would not have
+        # a chance to take any actions
+        nil
       end
 
       # Builds a new rdkafka consumer instance based on the subscription group configuration